Handle Google Sheets-style formula errors
Create a spreadsheet in both Excel and Google Sheets with nothing but a
formula, =FOO(), in cell A1. This will result in a formula error
(#NAME?) because there's no defined name FOO.

Save the file in Excel and download the sheet as an .xlsx from Google
Sheets. In the xl/worksheets/sheet1.xml file within the unzipped .xlsx
files, you'll see two different structures for the cell.

In Excel, it will stores the cell like this:

    <c r="A1" t="e" cm="1">
      <f t="array" aca="1" ref="A1" ca="1">FOO()</f>
      <v>#NAME?</v>
    </c>

In Google Sheets, it will store the cell like this:

    <c r="A1" s="1" t="str">
      <f>FOO()</f>
      <v>#NAME?</v>
    </c>

The key difference is that Excel marks the cell as having an error
(t="e", or type is error) but Google Sheets doesn't (t="str", or type is
string).

During conversion, xlsx-convert treats Excel's cell as an error, but
treats Google Sheets's cell as a plain string value. I have no idea if
Google Sheets's way of doing things is valid according to the spec, but
we should support these types of cells as errors to be consistent with
Google Sheets. Note that Excel does treat Google Sheets's cells as
errors.

This change checks whether a) a cell has a formula and b) the cell value
matches a known error string. If both are true, the cell is treated as
an error. If not, it's considered a string cell.

I've added two test spreadsheets: one with the error created in Excel,
one with the error created with Google Sheets.